Remove doc folder
The doc folder does not hold any easy to use and additional information. Moreless it convolutes the commits so that changes are not always visible at a glance.
Therefore it was decided to remove this folder and not keep it.

If you need to have a look at the usage, please refer to the code. Also the API intendent to be used is/will be briefly described in the Readme.md