Do asynchronous PV update in the initial context.
This is related to commit e90da91b3d9ddfec4cf2829014859f403d8787bb
The workaround there was to use a separate context for background
updates, which worked, but introduced the possibility of the two
contexts to be out of sync with regard to the connection status.

This commit reverts back to the original behaviour, i.e. the
background thread uses the same context (and PV objects) as the main
thread, eliminating potential issues. The original problem stems from
the fact that PvUpdater requested different data using ca.get() than
PV.get() did. When PV.get() tried to complete the request, it got data
that it didn't expect. The calls are now compatible.